Зробив перехід на myLibrary
але за замовчуванням не можу відобразити теку "watched"
