+++ src/index.html
@@ -22,7 +22,7 @@
 
     <script type="module" src="./scripts/gallery.js"></script>
     <script type="module" src="./scripts/header.js"></script>
-    <!-- <script type="module" src="./scripts/libraryBtnActive.js"></script> -->
+   
     <script type="module" src="./scripts/js-modal.js"></script>
     <script type="module" src="./scripts/pagination.js"></script>
     <script type="module" src="./js/theme-change.js"></script>
