+++ src/library.html
@@ -20,6 +20,15 @@
     <include src="./partials/theme-switcher.html"></include>
     <include src="./partials/footer.html"></include>
 
-  
+    <script type="module" src="./scripts/gallery.js"></script>
+   
+   
+    <script type="module" src="./scripts/js-modal.js"></script>
+    <script type="module" src="./scripts/pagination.js"></script>
+    <script type="module" src="./js/theme-change.js"></script>
+    <script type="module" src="./scripts/goToTop.js"></script>
+    <script type="module" src="./scripts/footer.js"></script>
+    <script type="module" src="./scripts/libraryBtnActive.js"></script>
+    
   </body>
 </html>
