+++ src/partials/gallery.html
@@ -1,9 +1,5 @@
 <section>
   <div class="container">
-    <button type="button" data-action="watched">watched</button>
-    <button type="button" data-action="queue">queue</button>
-    <button type="button" data-action="addWatched">addWatched</button>
-    <button type="button" data-action="addQueue">addQueue</button>
     <ul class="gallery"></ul>
     <div id="tui-pagination-container" class="tui-pagination"></div>
   </div>
