+++ src/partials/header-library.html
@@ -34,7 +34,7 @@
               HOME
             </a>
           </li>
-          <li class="header__nav-item">
+          <li class="header__nav-item library-item">
             <a href="./library.html" class="header__nav-link">MY LIBRARY</a>
           </li>
         </ul>
@@ -60,3 +60,11 @@
     </div>
   </div>
 </header>
+<!-- <section>
+  <div class="container">
+    
+    <ul class="galleryLib"></ul>
+    <div id="tui-pagination-container" class="tui-pagination"></div>
+  </div>
+  <a id="go-top"></a>
+</section> -->
