+++ src/partials/header.html
@@ -34,7 +34,7 @@
               HOME
             </a>
           </li>
-          <li class="header__nav-item">
+          <li class="header__nav-item library-item">
             <a href="./library.html" class="header__nav-link">MY LIBRARY</a>
           </li>
         </ul>
