+++ src/scripts/gallery.js
@@ -14,7 +14,8 @@ import {
 } from './api';
 
 import { Loading } from 'notiflix/build/notiflix-loading-aio';
-import { galleryRef } from './refs';
+import { galleryRef, formRef } from './refs';
+import { handleSearchFormSubmit } from './header';
 
 function markupRandomFilms() {
   fetchRandomFilm().then(({ results }) => 
@@ -61,75 +62,89 @@ export async function findGenres(genre_ids) {
   }
 }
 
+//=========== натискання на посилання MY LIBRARY =========================
+const myLibraryRef = document.querySelector('.library-item');
+console.log(myLibraryRef);
+
+function onMyLibraryLinkClick() {
+  console.log('клік працює');
+  formRef.removeEventListener('submit', handleSearchFormSubmit);
+
+  // galleryRef.innerHTML = '';
+  onWatchedBtnClick();
+} 
+
+myLibraryRef.addEventListener('click', onMyLibraryLinkClick);
+
 //============= Кнопка add to Watched ====================================
-const arrayFilmWatched = [];
-const addWatchedBtnRef = document.querySelector('button[data-action="addWatched"]');
+// const arrayFilmWatched = [];
+// const addWatchedBtnRef = document.querySelector('button[data-action="addWatched"]');
 
-async function onAddWatchedBtnClick() {
-  const randomId = Math.round(Math.random() * (100 - 1) + 1);
-  const value = await fetchFilmById(randomId);
-  const emptyValue = !Object.keys(value).length;
-  // const myValue = {[randomId]: value};
-  if (emptyValue) return;
-  arrayFilmWatched.push(value);
+// async function onAddWatchedBtnClick() {
+//   const randomId = Math.round(Math.random() * (100 - 1) + 1);
+//   const value = await fetchFilmById(randomId);
+//   const emptyValue = !Object.keys(value).length;
+//   // const myValue = {[randomId]: value};
+//   if (emptyValue) return;
+//   arrayFilmWatched.push(value);
   
-  saveToLS(LOCALSTORAGE_KEY_WATCHED, arrayFilmWatched);
-}
+//   saveToLS(LOCALSTORAGE_KEY_WATCHED, arrayFilmWatched);
+// }
 
-addWatchedBtnRef.addEventListener('click', onAddWatchedBtnClick);
+// addWatchedBtnRef.addEventListener('click', onAddWatchedBtnClick);
 
-async function fetchFilmById(movie_id) {
-  try {
-    const response = await axios.get(
-      `${BASE_URL}/movie/${movie_id}?api_key=${API_KEY}`
-    );
-    return response.data;
-  } catch (error) {
-    console.log(error);
-  }
-}
+// async function fetchFilmById(movie_id) {
+//   try {
+//     const response = await axios.get(
+//       `${BASE_URL}/movie/${movie_id}?api_key=${API_KEY}`
+//     );
+//     return response.data;
+//   } catch (error) {
+//     console.log(error);
+//   }
+// }
 
 //============= Кнопка Watched ====================================
-const LOCALSTORAGE_KEY_WATCHED = "watch";
-const watchedBtnRef = document.querySelector('button[data-action="watched"]');
+// const LOCALSTORAGE_KEY_WATCHED = "watch";
+// const watchedBtnRef = document.querySelector('button[data-action="watched"]');
 
-function onWatchedBtnClick() {
+// function onWatchedBtnClick() {
   
   
-  if (localStorage[LOCALSTORAGE_KEY_WATCHED]) {
-    const arrayFromLSWatch = loadFromLS(LOCALSTORAGE_KEY_WATCHED);
-    galleryRef.innerHTML = '';
-    const filmPromisesWatch = arrayFromLSWatch.map(id => fetchMovie(id));
+//   if (localStorage[LOCALSTORAGE_KEY_WATCHED]) {
+//     const arrayFromLSWatch = loadFromLS(LOCALSTORAGE_KEY_WATCHED);
+//     galleryRef.innerHTML = '';
+//     const filmPromisesWatch = arrayFromLSWatch.map(id => fetchMovie(id));
    
-    Promise.all(filmPromisesWatch).then(results =>        
-     results.map(
-     ({
-       poster_path,
-       backdrop_path,
-       original_title,
-       title,
-       genres,
-       release_date,
-       vote_average,
-       id,
-     }) => 
-       renderFilmList(
-         poster_path,
-         backdrop_path,
-         original_title,
-         title,
-         genres,
-         release_date,
-         vote_average,
-         id,
-         findGenres
-       )
-       ));
-   }
-    } 
+//     Promise.all(filmPromisesWatch).then(results =>        
+//      results.map(
+//      ({
+//        poster_path,
+//        backdrop_path,
+//        original_title,
+//        title,
+//        genres,
+//        release_date,
+//        vote_average,
+//        id,
+//      }) => 
+//        renderFilmList(
+//          poster_path,
+//          backdrop_path,
+//          original_title,
+//          title,
+//          genres,
+//          release_date,
+//          vote_average,
+//          id,
+//          findGenres
+//        )
+//        ));
+//    }
+//     } 
 
 
-watchedBtnRef.addEventListener('click', onWatchedBtnClick);
+// watchedBtnRef.addEventListener('click', onWatchedBtnClick);
 
 //============= Кнопка add to Queue ====================================
 
@@ -138,43 +153,43 @@ watchedBtnRef.addEventListener('click', onWatchedBtnClick);
 
 
 //============= Кнопка Queue ====================================
-const LOCALSTORAGE_KEY_QUEUE = "queue";
-const queueBtnRef = document.querySelector('button[data-action="queue"]');
+// const LOCALSTORAGE_KEY_QUEUE = "queue";
+// const queueBtnRef = document.querySelector('button[data-action="queue"]');
 
-function onQueueBtnClick() {
+// function onQueueBtnClick() {
   
-    if (localStorage[LOCALSTORAGE_KEY_QUEUE]) {
-       const arrayFromLSQueue = loadFromLS(LOCALSTORAGE_KEY_QUEUE);
-       galleryRef.innerHTML = '';
-       const filmPromisesQueue = arrayFromLSQueue.map(id => fetchMovie(id));
+//     if (localStorage[LOCALSTORAGE_KEY_QUEUE]) {
+//        const arrayFromLSQueue = loadFromLS(LOCALSTORAGE_KEY_QUEUE);
+//        galleryRef.innerHTML = '';
+//        const filmPromisesQueue = arrayFromLSQueue.map(id => fetchMovie(id));
       
-       Promise.all(filmPromisesQueue).then(results =>        
-        results.map(
-        ({
-          poster_path,
-          backdrop_path,
-          original_title,
-          title,
-          genres,
-          release_date,
-          vote_average,
-          id,
-        }) => 
-          renderFilmList(
-            poster_path,
-            backdrop_path,
-            original_title,
-            title,
-            genres,
-            release_date,
-            vote_average,
-            id,
-            findGenres
-          )
-          ));
-      }
-    } 
+//        Promise.all(filmPromisesQueue).then(results =>        
+//         results.map(
+//         ({
+//           poster_path,
+//           backdrop_path,
+//           original_title,
+//           title,
+//           genres,
+//           release_date,
+//           vote_average,
+//           id,
+//         }) => 
+//           renderFilmList(
+//             poster_path,
+//             backdrop_path,
+//             original_title,
+//             title,
+//             genres,
+//             release_date,
+//             vote_average,
+//             id,
+//             findGenres
+//           )
+//           ));
+//       }
+//     } 
 
 
-queueBtnRef.addEventListener('click', onQueueBtnClick);
+// queueBtnRef.addEventListener('click', onQueueBtnClick);
